+++ 2023/rust/src/template.rs
@@ -1,25 +1,20 @@
 use crate::parser;
 
-
-pub fn part1() -> u32 {
+pub fn part1() -> usize {
     0
 }
 
-pub fn part2() -> u32 {
+pub fn part2() -> usize {
     0
 }
-
 
 #[cfg(test)]
 mod tests {
     use super::*;
 
     #[test]
-    fn sample_input_part_1() {
-    }
+    fn sample_input_part_1() {}
 
     #[test]
-    fn sample_input_part_2() {
-    }
-
+    fn sample_input_part_2() {}
 }
